[75] Move home screen and viewmodel to feature package
